backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.22.0
pydantic==2.7.1
python-multipart==0.0.9

Flask==3.0.3
Flask-Cors==4.0.0

python-socketio==5.11.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4

requests==2.31.0
aiohttp==3.9.5
peewee==3.17.3
peewee-migrate==1.12.2
psycopg2-binary==2.9.9
PyMySQL==1.1.0
bcrypt==4.1.2

litellm[proxy]==1.35.28

boto3==1.34.95

argon2-cffi==23.1.0
APScheduler==3.10.4
google-generativeai==0.5.2

langchain==0.1.16
langchain-community==0.0.34
langchain-chroma==0.1.0

fake-useragent==1.5.1
chromadb==0.4.24
sentence-transformers==2.7.0
pypdf==4.2.0
docx2txt==0.8
python-pptx==0.6.23
unstructured==0.11.8
Markdown==3.6
pypandoc==1.13
pandas==2.2.2
openpyxl==3.1.2
pyxlsb==1.0.10
xlrd==2.0.1
validators==0.28.1

opencv-python-headless==4.9.0.80
rapidocr-onnxruntime==1.2.3

fpdf2==2.7.8
rank-bm25==0.2.2

faster-whisper==1.0.1

PyJWT[crypto]==2.8.0

black==24.4.2
langfuse==2.27.3
youtube-transcript-api==0.6.2
pytube
substrate-interface==1.7.8
web3
alibabacloud_tea_openapi==0.3.10
alibabacloud_cloudauth_intl20220809==2.0.2
# qrcode==7.4.2
alibabacloud-facebody20191230==5.1.2
oss2==2.18.6
aliyun-python-sdk-viapiutils==1.0
viapi-utils==1.0.1
tweepy==4.14.0
tavily_python==0.5.0
google-api-python-client==2.156.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
redis==5.2.1
jieba==0.42.1